My bad
(auto-msg whatthecommit.com)

On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	modified:   default.nix

bad things happen when you forget about that one little change you made ages ago

(auto-msg whatthecommit.com)

On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	modified:   default.nix

This Is Why We Don't Push To Production On Fridays

(auto-msg whatthecommit.com)

On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	modified:   default.nix